+++ src/Data/MemPack.hs
@@ -11,12 +11,13 @@
 {-# LANGUAGE NumericUnderscores #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE UnboxedTuples #-}
 
 -- |
 -- Module      : Data.MemPack
--- Copyright   : (c) Alexey Kuleshevich 2024
+-- Copyright   : (c) Alexey Kuleshevich 2024-2025
 -- License     : BSD3
 -- Maintainer  : Alexey Kuleshevich <alexey@kuleshevi.ch>
 -- Stability   : experimental
@@ -28,6 +29,7 @@ module Data.MemPack (
 
   -- * Packing
   pack,
+  packBuffer,
   packByteString,
   packShortByteString,
 
@@ -54,6 +56,8 @@ module Data.MemPack (
   unpackByteArrayLen,
   packByteStringM,
   unpackByteStringM,
+  packLiftST,
+  unpackLiftST,
 
   -- * Helper packers
   VarLen (..),
@@ -81,7 +85,7 @@ import Control.Monad (join, unless, when)
 import qualified Control.Monad.Fail as F
 import Control.Monad.Reader (MonadReader (..), lift)
 import Control.Monad.State.Strict (MonadState (..), StateT (..), execStateT)
-import Control.Monad.Trans.Fail (Fail, FailT (..), errorFail, failT, runFailAgg)
+import Control.Monad.Trans.Fail (Fail, FailT (..), errorFail, failT, runFailAgg, runFailAggT)
 import Data.Array.Byte (ByteArray (..), MutableByteArray (..))
 import Data.Bifunctor (first)
 import Data.Bits (Bits (..), FiniteBits (..))
@@ -91,11 +95,14 @@ import qualified Data.ByteString.Lazy.Internal as BSL
 import Data.ByteString.Short (ShortByteString)
 import Data.Char (ord)
 import Data.Complex (Complex (..))
+import qualified Data.Foldable as F (foldl')
 import Data.List (intercalate)
 import Data.MemPack.Buffer
 import Data.MemPack.Error
+import Data.Primitive.Array (Array (..), newArray, sizeofArray, unsafeFreezeArray, writeArray)
+import Data.Primitive.PrimArray (PrimArray (..), sizeofPrimArray)
+import Data.Primitive.Types (Prim (sizeOf#))
 import Data.Ratio
-import Data.Semigroup (Sum (..))
 #if MIN_VERSION_text(2,0,0)
 import qualified Data.Text.Array as T
 #endif
@@ -120,8 +127,11 @@ import GHC.Natural (Natural (..), isValidNatural)
 #else
 #error "Only integer-gmp is supported for now for older compilers"
 #endif
-#if !(MIN_VERSION_base(4,13,0))
+#if !MIN_VERSION_base(4,13,0)
 import Prelude (fail)
+#endif
+#if !MIN_VERSION_primitive(0,8,0)
+import qualified Data.Primitive.ByteArray as Prim (ByteArray(..))
 #endif
 
 -- | Monad that is used for serializing data into a `MutableByteArray`. It is based on
@@ -166,14 +176,14 @@ instance MonadState Int (Pack s) where
 -- `StateT` that tracks the current index into the @`Buffer` a@, from where the next read
 -- suppose to happen. Unpacking can `F.fail` with `F.MonadFail` instance or with
 -- `failUnpack` that provides a more type safe way of failing using `Error` interface.
-newtype Unpack b a = Unpack
-  { runUnpack :: b -> StateT Int (Fail SomeError) a
+newtype Unpack s b a = Unpack
+  { runUnpack :: b -> StateT Int (FailT SomeError (ST s)) a
   }
 
-instance Functor (Unpack s) where
+instance Functor (Unpack s b) where
   fmap f (Unpack p) = Unpack $ \buf -> fmap f (p buf)
   {-# INLINE fmap #-}
-instance Applicative (Unpack b) where
+instance Applicative (Unpack s b) where
   pure = Unpack . const . pure
   {-# INLINE pure #-}
   Unpack a1 <*> Unpack a2 =
@@ -182,23 +192,23 @@ instance Applicative (Unpack b) where
   Unpack a1 *> Unpack a2 =
     Unpack $ \buf -> a1 buf *> a2 buf
   {-# INLINE (*>) #-}
-instance Monad (Unpack b) where
+instance Monad (Unpack s b) where
   Unpack m1 >>= p =
     Unpack $ \buf -> m1 buf >>= \res -> runUnpack (p res) buf
   {-# INLINE (>>=) #-}
 #if !(MIN_VERSION_base(4,13,0))
   fail = Unpack . const . F.fail
 #endif
-instance F.MonadFail (Unpack b) where
+instance F.MonadFail (Unpack s b) where
   fail = Unpack . const . F.fail
-instance MonadReader b (Unpack b) where
+instance MonadReader b (Unpack s b) where
   ask = Unpack pure
   {-# INLINE ask #-}
   local f (Unpack p) = Unpack (p . f)
   {-# INLINE local #-}
   reader f = Unpack (pure . f)
   {-# INLINE reader #-}
-instance MonadState Int (Unpack b) where
+instance MonadState Int (Unpack s b) where
   get = Unpack $ const get
   {-# INLINE get #-}
   put = Unpack . const . put
@@ -206,7 +216,7 @@ instance MonadState Int (Unpack b) where
   state = Unpack . const . state
   {-# INLINE state #-}
 
-instance Alternative (Unpack b) where
+instance Alternative (Unpack s b) where
   empty = Unpack $ \_ -> lift empty
   {-# INLINE empty #-}
   Unpack r1 <|> Unpack r2 =
@@ -218,7 +228,7 @@ instance Alternative (Unpack b) where
   {-# INLINE (<|>) #-}
 
 -- | Failing unpacking with an `Error`.
-failUnpack :: Error e => e -> Unpack b a
+failUnpack :: Error e => e -> Unpack s b a
 failUnpack e = Unpack $ \_ -> lift $ failT (toSomeError e)
 
 -- | Efficient serialization interface that operates directly on memory buffers.
@@ -241,11 +251,13 @@ class MemPack a where
   packM :: a -> Pack s ()
 
   -- | Read binary representation of the type directly from the buffer, which can be
-  -- accessed with `ask` when necessary. Direct reads from the buffer should be preceded
-  -- with advancing the buffer offset with `MonadState` by the number of bytes that will
-  -- be consumed from the buffer and making sure that no reads outside of the buffer can
-  -- happen. Violation of these rules will lead to segfaults.
-  unpackM :: Buffer b => Unpack b a
+  -- accessed with `ask` when necessary.
+  --
+  -- /Warning/ - Direct reads from the buffer should be preceded with advancing the buffer offset
+  -- within `MonadState` by the exact number of bytes that gets consumed from that buffer.
+  --
+  -- __⚠__ - Violation of the above rule will lead to segfaults.
+  unpackM :: Buffer b => Unpack s b a
 
 instance MemPack () where
   packedByteCount _ = 0
@@ -320,7 +332,7 @@ instance MemPack Char where
     let c =
           buffer
             buf
-            (\ba# -> C# (indexWord8ArrayAsWideChar# ba# i#))
+            (\ba# off# -> C# (indexWord8ArrayAsWideChar# ba# (i# +# off#)))
             (\addr# -> C# (indexWideCharOffAddr# (addr# `plusAddr#` i#) 0#))
         ordc :: Word32
         ordc = fromIntegral (ord c)
@@ -344,7 +356,7 @@ instance MemPack Float where
     pure $!
       buffer
         buf
-        (\ba# -> F# (indexWord8ArrayAsFloat# ba# i#))
+        (\ba# off# -> F# (indexWord8ArrayAsFloat# ba# (i# +# off#)))
         (\addr# -> F# (indexFloatOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -362,7 +374,7 @@ instance MemPack Double where
     pure $!
       buffer
         buf
-        (\ba# -> D# (indexWord8ArrayAsDouble# ba# i#))
+        (\ba# off# -> D# (indexWord8ArrayAsDouble# ba# (i# +# off#)))
         (\addr# -> D# (indexDoubleOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -381,7 +393,7 @@ instance MemPack (Ptr a) where
     pure $!
       buffer
         buf
-        (\ba# -> Ptr (indexWord8ArrayAsAddr# ba# i#))
+        (\ba# off# -> Ptr (indexWord8ArrayAsAddr# ba# (i# +# off#)))
         (\addr# -> Ptr (indexAddrOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -400,7 +412,7 @@ instance MemPack (StablePtr a) where
     pure $!
       buffer
         buf
-        (\ba# -> StablePtr (indexWord8ArrayAsStablePtr# ba# i#))
+        (\ba# off# -> StablePtr (indexWord8ArrayAsStablePtr# ba# (i# +# off#)))
         (\addr# -> StablePtr (indexStablePtrOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -418,7 +430,7 @@ instance MemPack Int where
     pure $!
       buffer
         buf
-        (\ba# -> I# (indexWord8ArrayAsInt# ba# i#))
+        (\ba# off# -> I# (indexWord8ArrayAsInt# ba# (i# +# off#)))
         (\addr# -> I# (indexIntOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -436,7 +448,7 @@ instance MemPack Int8 where
     pure $!
       buffer
         buf
-        (\ba# -> I8# (indexInt8Array# ba# i#))
+        (\ba# off# -> I8# (indexInt8Array# ba# (i# +# off#)))
         (\addr# -> I8# (indexInt8OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -454,7 +466,7 @@ instance MemPack Int16 where
     pure $!
       buffer
         buf
-        (\ba# -> I16# (indexWord8ArrayAsInt16# ba# i#))
+        (\ba# off# -> I16# (indexWord8ArrayAsInt16# ba# (i# +# off#)))
         (\addr# -> I16# (indexInt16OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -472,7 +484,7 @@ instance MemPack Int32 where
     pure $!
       buffer
         buf
-        (\ba# -> I32# (indexWord8ArrayAsInt32# ba# i#))
+        (\ba# off# -> I32# (indexWord8ArrayAsInt32# ba# (i# +# off#)))
         (\addr# -> I32# (indexInt32OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -490,7 +502,7 @@ instance MemPack Int64 where
     pure $!
       buffer
         buf
-        (\ba# -> I64# (indexWord8ArrayAsInt64# ba# i#))
+        (\ba# off# -> I64# (indexWord8ArrayAsInt64# ba# (i# +# off#)))
         (\addr# -> I64# (indexInt64OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -508,7 +520,7 @@ instance MemPack Word where
     pure $!
       buffer
         buf
-        (\ba# -> W# (indexWord8ArrayAsWord# ba# i#))
+        (\ba# off# -> W# (indexWord8ArrayAsWord# ba# (i# +# off#)))
         (\addr# -> W# (indexWordOffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -526,7 +538,7 @@ instance MemPack Word8 where
     pure $!
       buffer
         buf
-        (\ba# -> W8# (indexWord8Array# ba# i#))
+        (\ba# off# -> W8# (indexWord8Array# ba# (i# +# off#)))
         (\addr# -> W8# (indexWord8OffAddr# addr# i#))
   {-# INLINE unpackM #-}
 
@@ -544,7 +556,7 @@ instance MemPack Word16 where
     pure $!
       buffer
         buf
-        (\ba# -> W16# (indexWord8ArrayAsWord16# ba# i#))
+        (\ba# off# -> W16# (indexWord8ArrayAsWord16# ba# (i# +# off#)))
         (\addr# -> W16# (indexWord16OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -562,7 +574,7 @@ instance MemPack Word32 where
     pure $!
       buffer
         buf
-        (\ba# -> W32# (indexWord8ArrayAsWord32# ba# i#))
+        (\ba# off# -> W32# (indexWord8ArrayAsWord32# ba# (i# +# off#)))
         (\addr# -> W32# (indexWord32OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -580,7 +592,7 @@ instance MemPack Word64 where
     pure $!
       buffer
         buf
-        (\ba# -> W64# (indexWord8ArrayAsWord64# ba# i#))
+        (\ba# off# -> W64# (indexWord8ArrayAsWord64# ba# (i# +# off#)))
         (\addr# -> W64# (indexWord64OffAddr# (addr# `plusAddr#` i#) 0#))
   {-# INLINE unpackM #-}
 
@@ -881,7 +893,12 @@ instance
 
 instance MemPack a => MemPack [a] where
   typeName = "[" ++ typeName @a ++ "]"
-  packedByteCount es = packedByteCount (Length (length es)) + getSum (foldMap (Sum . packedByteCount) es)
+  packedByteCount es =
+    let go [] (# listLen#, elemsLen# #) = packedByteCount (Length (I# listLen#)) + (I# elemsLen#)
+        go (x : xs) (# listLen#, elemsLen# #) =
+          let !(I# bc#) = packedByteCount x
+           in go xs (# 1# +# listLen#, bc# +# elemsLen# #)
+     in go es (# 0#, 0# #)
   {-# INLINE packedByteCount #-}
   packM as = do
     packM (Length (length as))
@@ -890,6 +907,27 @@ instance MemPack a => MemPack [a] where
   unpackM = do
     Length n <- unpackM
     replicateTailM n unpackM
+  {-# INLINE unpackM #-}
+
+instance MemPack a => MemPack (Array a) where
+  typeName = "(Array " ++ typeName @a ++ ")"
+  packedByteCount arr =
+    packedByteCount (Length (sizeofArray arr))
+      + F.foldl' (\acc e -> acc + packedByteCount e) 0 arr
+  {-# INLINE packedByteCount #-}
+  packM as = do
+    packM (Length (length as))
+    mapM_ packM as
+  {-# INLINE packM #-}
+  unpackM = do
+    Length n <- unpackM
+    marr <- unpackLiftST (newArray n (error "Uninitialized"))
+    let fill !i = when (i < n) $ do
+          e <- unpackM
+          unpackLiftST (writeArray marr i e)
+          fill (i + 1)
+    fill 0
+    unpackLiftST (unsafeFreezeArray marr)
   {-# INLINE unpackM #-}
 
 -- | Tail recursive version of `replicateM`
@@ -915,6 +953,38 @@ instance MemPack ByteArray where
   {-# INLINE packM #-}
   unpackM = unpackByteArray False
   {-# INLINE unpackM #-}
+
+instance (Typeable a, Prim a) => MemPack (PrimArray a) where
+  packedByteCount pa =
+    let len = I# (sizeOf# (undefined :: a)) * sizeofPrimArray pa
+     in packedByteCount (Length len) + len
+  {-# INLINE packedByteCount #-}
+  packM pa@(PrimArray ba#) = do
+    let !len@(I# len#) = I# (sizeOf# (undefined :: a)) * sizeofPrimArray pa
+    packM (Length len)
+    I# curPos# <- state $ \i -> (i, i + len)
+    MutableByteArray mba# <- ask
+    lift_# (copyByteArray# ba# 0# mba# curPos# len#)
+  {-# INLINE packM #-}
+  unpackM = (\(ByteArray ba#) -> PrimArray ba#) <$> unpackByteArray False
+  {-# INLINE unpackM #-}
+
+#if !MIN_VERSION_primitive(0,8,0)
+instance MemPack Prim.ByteArray where
+  packedByteCount ba =
+    let len = bufferByteCount ba
+     in packedByteCount (Length len) + len
+  {-# INLINE packedByteCount #-}
+  packM ba@(Prim.ByteArray ba#) = do
+    let !len@(I# len#) = bufferByteCount ba
+    packM (Length len)
+    I# curPos# <- state $ \i -> (i, i + len)
+    MutableByteArray mba# <- ask
+    lift_# (copyByteArray# ba# 0# mba# curPos# len#)
+  {-# INLINE packM #-}
+  unpackM = (\(ByteArray ba#) -> Prim.ByteArray ba#) <$> unpackByteArray False
+  {-# INLINE unpackM #-}
+#endif
 
 instance MemPack ShortByteString where
   packedByteCount ba =
@@ -983,7 +1053,9 @@ instance MemPack Text where
     MutableByteArray mba# <- ask
     lift_# (copyByteArray# ba# offset# mba# curPos# len#)
 #else
-  -- FIXME: This is very inefficient and shall be fixed in the next major version
+  -- FIXME: This is very inefficient and hopefully will be fixed at some point.  It requires some
+  -- clever change to the MemPack interface in order to allow memoization between `packedByteCount`
+  -- and `packM`
   packedByteCount = packedByteCount . T.encodeUtf8
   packM = packM . T.encodeUtf8
 #endif
@@ -998,7 +1070,7 @@ instance MemPack Text where
 {- FOURMOLU_ENABLE -}
 
 -- | This is the implementation of `unpackM` for `ByteArray`, `ByteString` and `ShortByteString`
-unpackByteArray :: Buffer b => Bool -> Unpack b ByteArray
+unpackByteArray :: Buffer b => Bool -> Unpack s b ByteArray
 unpackByteArray isPinned = unpackByteArrayLen isPinned . unLength =<< unpackM
 {-# INLINE unpackByteArray #-}
 
@@ -1007,7 +1079,7 @@ unpackByteArray isPinned = unpackByteArrayLen isPinned . unLength =<< unpackM
 -- Similar to `unpackByteArray`, except it does not unpack a length.
 --
 -- @since 0.1.1
-unpackByteArrayLen :: Buffer b => Bool -> Int -> Unpack b ByteArray
+unpackByteArrayLen :: Buffer b => Bool -> Int -> Unpack s b ByteArray
 unpackByteArrayLen isPinned len@(I# len#) = do
   I# curPos# <- guardAdvanceUnpack len
   buf <- ask
@@ -1015,7 +1087,7 @@ unpackByteArrayLen isPinned len@(I# len#) = do
     mba@(MutableByteArray mba#) <- newMutableByteArray isPinned len
     buffer
       buf
-      (\ba# -> st_ (copyByteArray# ba# curPos# mba# 0# len#))
+      (\ba# off# -> st_ (copyByteArray# ba# (curPos# +# off#) mba# 0# len#))
       (\addr# -> st_ (copyAddrToByteArray# (addr# `plusAddr#` curPos#) mba# 0# len#))
     freezeMutableByteArray mba
 {-# INLINE unpackByteArrayLen #-}
@@ -1032,7 +1104,7 @@ packIncrement a =
 -- | Increment the offset counter of `Unpack` monad by the supplied number of
 -- bytes. Returns the original offset or fails with `RanOutOfBytesError` whenever there is
 -- not enough bytes in the `Buffer`.
-guardAdvanceUnpack :: Buffer b => Int -> Unpack b Int
+guardAdvanceUnpack :: Buffer b => Int -> Unpack s b Int
 guardAdvanceUnpack n@(I# n#) = do
   buf <- ask
   let !len = bufferByteCount buf
@@ -1044,7 +1116,7 @@ guardAdvanceUnpack n@(I# n#) = do
       _ -> (failOutOfBytes i len n, i)
 {-# INLINE guardAdvanceUnpack #-}
 
-failOutOfBytes :: Int -> Int -> Int -> Unpack b a
+failOutOfBytes :: Int -> Int -> Int -> Unpack s b a
 failOutOfBytes i len n =
   failUnpack $
     toSomeError $
@@ -1070,6 +1142,27 @@ pack = packByteArray False
 packByteString :: forall a. (MemPack a, HasCallStack) => a -> ByteString
 packByteString = pinnedByteArrayToByteString . packByteArray True
 {-# INLINE packByteString #-}
+
+-- | Serialize a type into any `Buffer`
+--
+-- prop> pack xs == packBuffer xs
+--
+-- prop> packByteString xs == packBuffer xs
+--
+-- ====__Examples__
+--
+-- >>> :set -XTypeApplications
+-- >>> import qualified Data.Vector.Primitive as VP
+-- >>> import Data.Word (Word8)
+-- >>> unpack @[Int] $ packBuffer @[Int] @(VP.Vector Word8) [1,2,3,4,5]
+-- Right [1,2,3,4,5]
+--
+-- @since 0.2.0
+packBuffer :: forall a b. (MemPack a, Buffer b, HasCallStack) => a -> b
+packBuffer a =
+  case packByteArray (bufferHasToBePinned @b) a of
+    ByteArray ba -> mkBuffer ba
+{-# INLINE packBuffer #-}
 
 -- | Serialize a type into an unpinned `ShortByteString`
 packShortByteString :: forall a. (MemPack a, HasCallStack) => a -> ShortByteString
@@ -1160,7 +1253,7 @@ packByteStringM :: ByteString -> Pack s ()
 packByteStringM bs = do
   let !len@(I# len#) = bufferByteCount bs
   I# curPos# <- state $ \i -> (i, i + len)
-  Pack $ \(MutableByteArray mba#) -> lift $ withPtrByteStringST bs $ \(Ptr addr#) ->
+  Pack $ \(MutableByteArray mba#) -> lift $ withAddrByteStringST bs $ \addr# ->
     st_ (copyAddrToByteArray# addr# mba# curPos# len#)
 {-# INLINE packByteStringM #-}
 
@@ -1171,19 +1264,26 @@ unpackByteStringM ::
   Buffer b =>
   -- | number of bytes to unpack
   Int ->
-  Unpack b ByteString
+  Unpack s b ByteString
 unpackByteStringM len = pinnedByteArrayToByteString <$> unpackByteArrayLen True len
 {-# INLINE unpackByteStringM #-}
 
 -- | Unpack a memory `Buffer` into a type using its `MemPack` instance. Besides the
 -- unpacked type it also returns an index into a buffer where unpacked has stopped.
 unpackLeftOver :: forall a b. (MemPack a, Buffer b, HasCallStack) => b -> Fail SomeError (a, Int)
-unpackLeftOver b = do
+unpackLeftOver b = FailT $ pure $ runST $ runFailAggT $ unpackLeftOverST b
+{-# INLINE unpackLeftOver #-}
+
+-- | Unpack a memory `Buffer` into a type using its `MemPack` instance. Besides the
+-- unpacked type it also returns an index into a buffer where unpacked has stopped.
+unpackLeftOverST ::
+  forall a b s. (MemPack a, Buffer b, HasCallStack) => b -> FailT SomeError (ST s) (a, Int)
+unpackLeftOverST b = do
   let len = bufferByteCount b
   res@(_, consumedBytes) <- runStateT (runUnpack unpackM b) 0
   when (consumedBytes > len) $ errorLeftOver (typeName @a) consumedBytes len
   pure res
-{-# INLINEABLE unpackLeftOver #-}
+{-# INLINEABLE unpackLeftOverST #-}
 
 -- | This is a critical error, therefore we are not gracefully failing this unpacking
 errorLeftOver :: HasCallStack => String -> Int -> Int -> a
@@ -1348,13 +1448,13 @@ packIntoCont7 x cont n
 unpack7BitVarLen ::
   (Num a, Bits a, Buffer b) =>
   -- | Continuation that will be invoked if MSB is set
-  (Word8 -> a -> Unpack b a) ->
+  (Word8 -> a -> Unpack s b a) ->
   -- | Will be set either to 0 initially or to the very first unmodified byte, which is
   -- guaranteed to have the first bit set.
   Word8 ->
   -- | Accumulator
   a ->
-  Unpack b a
+  Unpack s b a
 unpack7BitVarLen cont firstByte !acc = do
   b8 :: Word8 <- unpackM
   if b8 `testBit` 7
@@ -1364,12 +1464,12 @@ unpack7BitVarLen cont firstByte !acc = do
 {-# INLINE unpack7BitVarLen #-}
 
 unpack7BitVarLenLast ::
-  forall t b.
+  forall t b s.
   (Num t, Bits t, MemPack t, Buffer b) =>
   Word8 ->
   Word8 ->
   t ->
-  Unpack b t
+  Unpack s b t
 unpack7BitVarLenLast mask firstByte acc = do
   res <- unpack7BitVarLen (\_ _ -> F.fail "Too many bytes.") firstByte acc
   -- Only while decoding the last 7bits we check if there was too many
@@ -1440,7 +1540,7 @@ packedTagByteCount :: Int
 packedTagByteCount = SIZEOF_WORD8
 {-# INLINE packedTagByteCount #-}
 
-unpackTagM :: Buffer b => Unpack b Tag
+unpackTagM :: Buffer b => Unpack s b Tag
 unpackTagM = Tag <$> unpackM
 {-# INLINE unpackTagM #-}
 
@@ -1458,3 +1558,17 @@ lift_# f = Pack $ \_ -> lift $ st_ f
 st_ :: (State# s -> State# s) -> ST s ()
 st_ f = ST $ \s# -> (# f s#, () #)
 {-# INLINE st_ #-}
+
+-- | Lift an `ST` action into the `Pack` monad
+--
+-- @since 0.2.0
+packLiftST :: ST s a -> Pack s a
+packLiftST st = Pack (\_ -> StateT (\i -> (,i) <$> st))
+{-# INLINE packLiftST #-}
+
+-- | Lift an `ST` action into the `Unpack` monad
+--
+-- @since 0.2.0
+unpackLiftST :: ST s a -> Unpack s b a
+unpackLiftST st = Unpack (\_ -> StateT (\i -> FailT (Right . (,i) <$> st)))
+{-# INLINE unpackLiftST #-}